requirements.txt
PyQt==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.2

numpy==2.2.6
rasterio==1.4.3
matplotlib==3.10.3

